code clean up, moved game loop to main.cpp
window.cpp was getting too cluttered, so moved game loop and destructor calls into main.cpp, game loop handling was not the responsibility of the window anyways
